requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
chromadb
pypdf
sentence-transformers
python-dotenv